Make `ChSuspension` the base class for all suspension subsystems.
With this, `ChDoubleWishboneReduced` becomes an actual suspension subsystem template (one of several possible) which describes the general topology and functionality of a particular type of suspension (in this case, a double a-arm suspension modeled with constraints).
In turn, `HMMWV9_DoubleWishboneFront` and `HMMWV9_DoubleWishboneRear` are concrete instantiations of that template (using particular values appropriate for the HMMWV front and rear suspensions, respectively).