Modify the mechanism to attach a wheel to a suspension subsystem.
Now, this is done in the wheel's Initialize() function (instead of calling the suspension's AttachWheel() function).
The base class implementation ChWheel::Initialize() does the work of incrementing the spindle's mass and inertia with that of the wheel.
A derived class must first call the base method before doing any other initializations.