Change signature of ChPacejkaTire constructor.
The alternate constructor (used for testing/debugging) must always specifies the tire vertical load (no default provided).  However, the caller can still pass a negative value, in which case the vertical load is internally calculated.
Note: this change is required in order to remove the ambiguity between the two constructors.